refactored interfaces
