--- src/components/Common/Divider/Divider.spec.ts
@@ -1,3 +1,0 @@
-export interface DividerSpec {
-    className?: string,
-}
