--- src/components/Common/ProfileSelector/ProfileSelector.test.ts
@@ -1,7 +1,0 @@
-describe('ProfileSelector', () => {
-	it('Should be implemented', () => {
-		console.log('Todo');
-	})
-});
-
-export {}
