--- src/components/Common/SectionTitle/SectionTitle.spec.ts
@@ -1,8 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface SectionTitleSpec {
-	children: ReactNode;
-	subtitle?: string;
-	hintColor?: string;
-	width?: string;
-}
