--- src/components/Common/Title/Title.spec.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface TitleSpec {
-	children: ReactNode;
-}
