--- src/components/Homepage/RuleItem/RuleItem.spec.ts
@@ -1,5 +1,0 @@
-export interface RuleItemSpec {
-	id: number,
-	title: string,
-	content?: string,
-}
