--- src/components/Lobby/GameView/UserProfile.spec.ts
@@ -1,7 +1,0 @@
-import Player from '../../../../server/classes/Player';
-
-export interface UserProfileSpec {
-	player: Player;
-	creatorId: string;
-	currentPlayer: Player;
-}
