--- src/components/Lobby/LobbyView/UserCard.spec.ts
@@ -1,8 +1,0 @@
-import Player from '../../../../server/classes/Player';
-
-export interface UserCardSpec {
-	player: Player;
-	creatorId?: string;
-	currentPlayerId?: string | unknown;
-	subtitle?: string;
-}
